Do not attempt to kill PID 0
If an extension dev happens to be manually running the sidecar in
'dev' mode, it won't have done its full initialization and may
return 401 from the health check route, but claiming to be
PID 0 in the response headers.

Guard against that by noticing if <= 0, then not raising the
exception that will lead us to try to kill().