Minor cleanups. (#302)
* When asking the topics view panel to reset due to a known change to
the topics in a cluster, provide the cluster id to the reset() call so
that the view panel can filter out repaint requests if not viewing said
cluster's topics.

  * No need for data member TopicViewProvider.ccloudEnvironment, can be
simplified away since its primary use we have a local var that's just
as good. One fewer moving part to track then.