Fix order of operations within ccloudProvider::getSessions() (#2182)
Need to update the context value governing ccloud-connectedness *before*
firing off the ccloudConnected event, because many event handlers
there ignore the boolean passed into the event emitter and instead
check hasCCloudAuthSession() which is based on the context value.

The issue was identified while working on the new resources
panel, but is fixable independently in a small branch.