fix: Don't clear diagnostics when saving documents (#2354)
We clear diagnostics in the listener for document change events. Prior to this change, we cleared diagnostics upon saving documents because saving a document generates a document change event.

This change updates the event listener, so that we clear diagnostics only if a document change event updated the content of the document, which is not the case when saving it.

Fixes #2350