[main] pull in fixes from `v1.5.x` (#2358)
* fix: Don't clear diagnostics when saving documents (#2354)

We clear diagnostics in the listener for document change events. Prior to this change, we cleared diagnostics upon saving documents because saving a document generates a document change event.

This change updates the event listener, so that we clear diagnostics only if a document change event updated the content of the document, which is not the case when saving it.

Fixes #2350

* [v1.5.x] Trigger diagnostics when opening document (#2348)

* fix: Trigger diagnostics when opening document

Upon opening a new document, send the `textDocument/didChange` event to the server
to trigger diagnostics. This change can be reverted once the CCloud Flink Language Server
has been updated to run diagnostics on the `textDocument/didOpen` event.

Fixes #2343

* chore: Improve function signature

* chore: Argh, adding missing semicolon

* chore: Incorporate reviewer feedback

* chore: Trigger diagnostics after attempting language client start (#2353)

After attempting a language client start, always trigger diagnostics
by simulating a change to the active document.

Fixes #2349

* fix test; add new test

* update comment

* add tests for `simulateDocumentChangeToTriggerDiagnostics()`

* add tests for `maybeStartLanguageClient()`

---------

Co-authored-by: Stefan Sprenger <128683+flippingbits@users.noreply.github.com>